Refactored folder & file structure
