+++ app/styles/RuleList.module.scss
