+++ app/styles/SectionList.module.scss
