+++ app/styles/SubruleList.module.scss
@@ -2,6 +2,6 @@
   font-size: 12px;
 }
 
-.example {
+// .example {
 
-}
+// }
