+++ app/styles/TocChapterList.module.scss
@@ -1,0 +1,3 @@
+// .tocChapters {
+//   // width: 450px;
+// }
