+++ pages/_app.js
@@ -1,4 +1,4 @@
-import '../styles/globals.scss';
+import '../app/styles/globals.scss';
 import 'bootstrap/dist/css/bootstrap.min.css';
