--- styles/TocChapterList.module.scss
@@ -1,3 +1,0 @@
-.tocChapters {
-  // width: 450px;
-}
